+++ Editor/Editor.vcxproj.filters
@@ -168,6 +168,7 @@
     <ClCompile Include="LightMeshGenerator.cpp">
       <Filter>Entities\Light</Filter>
     </ClCompile>
+    <ClCompile Include="DynamicMenu.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Mod.h">
@@ -298,6 +299,9 @@
     </ClInclude>
     <ClInclude Include="LightMeshGenerator.h">
       <Filter>Entities\Light</Filter>
+    </ClInclude>
+    <ClInclude Include="DynamicMenu.h">
+      <Filter>UI</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
